2019-01bW.p
Sorry for interfering Patrick, but I couldn't help editing your source. I only added a variable for the Fuel and in the repeat loop I calculate the fuel from the mass and after that set the fuel as the new mass (until the fuel required is less than or equal to zero.
I think it makes the program more readable.